[pragma]: Added suuport for pragma unordered_for
In this commit I have added the pragma for unordered_for. This
pragma can be applied to any level of loop nest and can be verified
that it is working by doing:

% make clang

in the tests directory of the main repo. This will generate a .ll
file and you should see that !{!"llvm.loop.unordered_for.enable"}
gets applied to one loop